Make version visible to users Fixes #6
