+++ migration.go
@@ -1,22 +1,5 @@
 package migrator
 
-import (
-	"os"
-
-	"gopkg.in/yaml.v3"
-)
-
-// Migrations is a collection of Migrations, i.e. the YAML file.
-type Migrations struct {
-	Migrations []Migration `yaml:"migrations"`
-}
-
-func (ms Migrations) enumerateMigrations() {
-	for i := range ms.Migrations {
-		ms.Migrations[i].version = i + 1
-	}
-}
-
 // Migration represents an entry defined in the migration YAML.
 type Migration struct {
 	Comment string `yaml:"comment"`
@@ -26,20 +9,10 @@ type Migration struct {
 	version int    `yaml:"-"`
 }
 
-func load() (Migrations, error) {
-	filename, found := os.LookupEnv(envVarFile)
-	if !found {
-		return Migrations{}, ErrMigrationFileEnvMissing
-	}
-	b, err := os.ReadFile(filename)
-	if err != nil {
-		return Migrations{}, err
-	}
-	migrations := Migrations{}
-	if err := yaml.Unmarshal(b, &migrations); err != nil {
-		return Migrations{}, err
-	}
-	return migrations, nil
+// Version return the version number given for this migration. A migration gets
+// it version from its position in the migrations YAML-file.
+func (m Migration) Version() int {
+	return m.version
 }
 
 func (m Migration) stmt(dir direction) string {
